Most important changes here - I moved the Box Mode, viewport tags, and inline classes into _framework, since ideally these will never need to be touched. On the other side of the dime, _base will be second least-used (I want most of the settings to take place in _siteSettings), while _custom and _print are going to be the heavy hitters.
Also of note: I now have a _siteSettings file as a direct overwrite of Foundation's _settings. _settings is still present for defaults and convenience - however, I aim for _siteSettings to be independent of any changes that Foundation may implement, as updating with every version can be a nuisance, and I want to minimize how much I need to change. (I may rename _settings to better indicate that it's not the focal point.)